refactor: extract tool/curl/sql presentation helpers from standalone-chat (3/N)
Move the ToolCall-free string/URL/SQL command parsers (classifyCurl,
endpointFamily, summarizeToolResult, formatToolResult, sanitizeCommand, the
curl/url/sql helpers + their types) into lib/chat/tool-presentation.ts. Pure
code-move, no behavior change. The ToolCall-typed extractors stay in place for
now (they depend on the shared ToolCall type — a later PR).

Adds 29 characterization tests covering classifyCurl's endpoint routing, SQL
verb/table parsing, curl body/method parsing, result summarization, and token
redaction.

tsc --noEmit: 0 errors. vitest: 408/408 (incl. 29 new). standalone-chat.tsx
9642 -> 9159 lines.

Follow-up to maintainer request:
https://github.com/screenpipe/screenpipe/pull/3754#issuecomment-4594997031
