+++ apps/screenpipe-app-tauri/components/standalone-chat.tsx
@@ -103,6 +103,21 @@ import {
   PI_MAX_RATE_LIMIT_RETRIES,
 } from "@/lib/chat/quota-errors";
 import { buildSystemPrompt, buildConnectionsContext } from "@/lib/chat/system-prompt";
+import {
+  classifyCurl,
+  endpointFamily,
+  summarizeToolResult,
+  formatToolResult,
+  sanitizeCommand,
+  curlBodyJson,
+  curlMethod,
+  trunc,
+  sqlTables,
+  urlsInCommand,
+  isLocalScreenpipeUrl,
+  firstExternalWebTarget,
+  type WebTargetPresentation,
+} from "@/lib/chat/tool-presentation";
 import { usePipes } from "@/lib/hooks/use-pipes";
 import { localFetch, getApiBaseUrl } from "@/lib/api";
 import { CONNECTIONS_UPDATED_EVENT } from "@/lib/connections-events";
@@ -772,391 +787,6 @@ function GridDissolveLoader({
 // can show "Searched ChatGPT 'foo'" instead of the raw curl URL. Pi's pipes
 // emit these as plain bash tool calls (no MCP), with the app name encoded as
 // app_name=X in the query string — see crates/screenpipe-core/assets/pipes/.
-interface SearchInfo {
-  appName?: string;
-  windowName?: string;
-  query?: string;
-  contentType?: string;
-}
-function parseSearchCommand(cmd: string): SearchInfo | null {
-  if (!cmd) return null;
-  const m = cmd.match(/https?:\/\/[^\s'"`]+\/search\?[^\s'"`]+/);
-  if (!m) return null;
-  try {
-    const url = new URL(m[0]);
-    if (!url.pathname.endsWith("/search")) return null;
-    const sp = url.searchParams;
-    const out: SearchInfo = {
-      appName: sp.get("app_name") || undefined,
-      windowName: sp.get("window_name") || undefined,
-      query: sp.get("q") || undefined,
-      contentType: sp.get("content_type") || undefined,
-    };
-    if (!out.appName && !out.windowName && !out.query && !out.contentType) return null;
-    return out;
-  } catch {
-    return null;
-  }
-}
-
-// Reads the JSON payload from a `-d '<json>'` (or --data / --data-raw)
-// argument of a curl command. Single-quoted is the common shape in pi's
-// pipes; double-quoted with backslash-escaped inner quotes is the fallback.
-function curlBodyJson(cmd: string): any | null {
-  let m = cmd.match(/(?:-d|--data(?:-raw|-binary)?)\s+'((?:[^'\\]|\\.)*)'/s);
-  let raw = m ? m[1] : null;
-  if (!raw) {
-    m = cmd.match(/(?:-d|--data(?:-raw|-binary)?)\s+"((?:[^"\\]|\\.)*)"/s);
-    raw = m ? m[1].replace(/\\"/g, '"') : null;
-  }
-  if (!raw) return null;
-  try { return JSON.parse(raw); } catch {}
-  try { return JSON.parse(raw.replace(/\\'/g, "'")); } catch {}
-  return null;
-}
-
-function curlMethod(cmd: string): string {
-  if (/(^|\s)(?:-I|--head)(?=\s|$)/i.test(cmd)) return "HEAD";
-  const m = cmd.match(/(?:-X|--request)\s+([A-Z]+)/i);
-  return m ? m[1].toUpperCase() : "GET";
-}
-
-function trunc(s: string, n: number): string {
-  return s.length > n ? s.slice(0, n) + "…" : s;
-}
-
-function sqlTables(sql: string): string[] {
-  const out = new Set<string>();
-  const re = /(?:FROM|JOIN)\s+([a-zA-Z_][a-zA-Z0-9_]*)/gi;
-  let m: RegExpExecArray | null;
-  while ((m = re.exec(sql)) !== null) out.add(m[1].toLowerCase());
-  return Array.from(out);
-}
-
-function sqlVerb(sql: string): string {
-  const s = sql.trim().toUpperCase();
-  if (s.startsWith("SELECT")) {
-    if (/^\s*SELECT\s+COUNT\s*\(/i.test(sql.trim())) return "Counted";
-    return "Queried";
-  }
-  if (s.startsWith("WITH")) return "Queried";
-  if (s.startsWith("INSERT")) return "Inserted into";
-  if (s.startsWith("UPDATE")) return "Updated";
-  if (s.startsWith("DELETE")) return "Deleted from";
-  return "Ran SQL on";
-}
-
-type WebTargetKind = "fetch" | "navigate" | "eval";
-
-interface WebTargetPresentation {
-  url: string;
-  domain: string;
-  label: string;
-  kind: WebTargetKind;
-}
-
-interface CurlPresentation {
-  label: string;
-  appName?: string;
-  connectionIconName?: string;
-  webTarget?: WebTargetPresentation;
-}
-
-function parseUrlCandidate(raw: string): URL | null {
-  let candidate = raw;
-  for (let i = 0; i < 4; i++) {
-    try {
-      return new URL(candidate);
-    } catch {
-      candidate = candidate.replace(/[),.;\]}]+$/, "");
-    }
-  }
-  return null;
-}
-
-function urlsInCommand(cmd: string): URL[] {
-  return Array.from(cmd.matchAll(/https?:\/\/[^\s'"`<>]+/g))
-    .map((m) => parseUrlCandidate(m[0]))
-    .filter((url): url is URL => Boolean(url));
-}
-
-function isLocalScreenpipeUrl(url: URL): boolean {
-  return (url.hostname === "localhost" || url.hostname === "127.0.0.1") && url.port === "3030";
-}
-
-function domainForUrl(url: URL): string {
-  return url.hostname.replace(/^www\./i, "");
-}
-
-function displayWebUrl(url: URL): string {
-  const domain = domainForUrl(url);
-  const path = `${url.pathname}${url.search}`;
-  return path && path !== "/" ? trunc(`${domain}${path}`, 48) : domain;
-}
-
-function webTargetFromUrl(url: URL, kind: WebTargetKind): WebTargetPresentation | null {
-  if (isLocalScreenpipeUrl(url)) return null;
-  return {
-    url: url.toString(),
-    domain: domainForUrl(url),
-    label: displayWebUrl(url),
-    kind,
-  };
-}
-
-function webTargetFromUrlString(raw: string, kind: WebTargetKind): WebTargetPresentation | null {
-  const url = parseUrlCandidate(raw);
-  return url ? webTargetFromUrl(url, kind) : null;
-}
-
-function firstExternalWebTarget(cmd: string, kind: WebTargetKind): WebTargetPresentation | null {
-  for (const url of urlsInCommand(cmd)) {
-    const target = webTargetFromUrl(url, kind);
-    if (target) return target;
-  }
-  return null;
-}
-
-function externalCurlLabel(method: string, target: WebTargetPresentation): string {
-  if (method === "GET") return `Fetched ${target.domain}`;
-  if (method === "HEAD") return `Checked ${target.domain}`;
-  if (method === "POST") return `Posted to ${target.domain}`;
-  return `${method} ${target.domain}`;
-}
-
-// Maps pi's bash curl calls to the local screenpipe API into a human label.
-// Pi only emits raw curl (no MCP), so the action lives in the URL + body —
-// every endpoint family pi actually uses (sampled from ~/.pi/agent/sessions)
-// gets its own rewrite here. Unknown endpoints fall back to "<METHOD> <path>".
-function classifyCurl(cmd: string): CurlPresentation | null {
-  if (!cmd) return null;
-
-  const search = parseSearchCommand(cmd);
-  if (search) {
-    const target = search.appName || search.windowName || search.contentType || "recordings";
-    const q = search.query ? ` "${trunc(search.query, 40)}"` : "";
-    return { label: `Searched ${target}${q}`, appName: search.appName || search.windowName };
-  }
-
-  const method = curlMethod(cmd);
-  const urls = urlsInCommand(cmd);
-  const url = urls.find(isLocalScreenpipeUrl);
-  if (!url) {
-    const target = firstExternalWebTarget(cmd, "fetch");
-    if (!target || !/\bcurl\b/i.test(cmd)) return null;
-    return { label: externalCurlLabel(method, target), webTarget: target };
-  }
-
-  const path = url.pathname.replace(/\/$/, "") || "/";
-
-  if (path === "/raw_sql") {
-    const body = curlBodyJson(cmd);
-    const sql = body && typeof body.query === "string" ? body.query : null;
-    if (!sql) return { label: "Ran SQL" };
-    const tables = sqlTables(sql);
-    const verb = sqlVerb(sql);
-    if (tables.length === 0) return { label: verb };
-    if (tables.length === 1) return { label: `${verb} ${tables[0]}` };
-    return { label: `${verb} ${tables.slice(0, 2).join(" + ")}${tables.length > 2 ? " +…" : ""}` };
-  }
-
-  if (path === "/activity-summary") return { label: "Activity summary" };
-
-  if (path === "/memories") {
-    if (method === "POST") return { label: "Saved memory" };
-    return { label: "Listed memories" };
-  }
-  const memMatch = path.match(/^\/memories\/(\w+)$/);
-  if (memMatch) {
-    const id = memMatch[1];
-    if (method === "PATCH" || method === "PUT") return { label: `Updated memory #${id}` };
-    if (method === "DELETE") return { label: `Deleted memory #${id}` };
-    return { label: `Got memory #${id}` };
-  }
-
-  if (path === "/meetings") return { label: "Listed meetings" };
-  const meetingMatch = path.match(/^\/meetings\/(\w+)$/);
-  if (meetingMatch) {
-    const id = meetingMatch[1];
-    if (method === "PATCH" || method === "PUT") {
-      const body = curlBodyJson(cmd);
-      const hasTitle = body && typeof body.title === "string" && body.title.length > 0;
-      const hasNote = body && typeof body.note === "string" && body.note.length > 0;
-      if (hasTitle && hasNote) return { label: `Renamed + summarized meeting #${id}` };
-      if (hasNote) return { label: `Summarized meeting #${id}` };
-      if (hasTitle) return { label: `Renamed meeting #${id}` };
-      return { label: `Updated meeting #${id}` };
-    }
-    if (method === "DELETE") return { label: `Deleted meeting #${id}` };
-    return { label: `Got meeting #${id}` };
-  }
-
-  if (path === "/speakers/similar") {
-    const name = url.searchParams.get("name") || url.searchParams.get("speaker_name");
-    return { label: name ? `Found similar speakers for "${trunc(name, 30)}"` : "Found similar speakers" };
-  }
-  if (path === "/speakers/merge") return { label: "Merged speakers" };
-  if (path === "/speakers/search") {
-    const q = url.searchParams.get("name") || url.searchParams.get("q");
-    return { label: q ? `Searched speakers "${trunc(q, 30)}"` : "Searched speakers" };
-  }
-  if (path === "/speakers/unnamed") return { label: "Listed unnamed speakers" };
-  if (path.startsWith("/speakers/")) {
-    const id = path.split("/")[2];
-    if (method === "PATCH" || method === "PUT") return { label: `Renamed speaker #${id}` };
-    return { label: `Got speaker #${id}` };
-  }
-
-  if (path === "/connections/browsers/owned-default/navigate") {
-    const body = curlBodyJson(cmd);
-    if (body && typeof body.url === "string") {
-      const target = webTargetFromUrlString(body.url, "navigate");
-      if (target) return { label: `Opened ${target.domain} in agent browser`, webTarget: target };
-    }
-    return { label: "Navigated agent browser" };
-  }
-  if (path === "/connections/browsers/owned-default/eval") {
-    const body = curlBodyJson(cmd);
-    if (body && typeof body.url === "string") {
-      const target = webTargetFromUrlString(body.url, "eval");
-      if (target) return { label: `Ran JS on ${target.domain}`, webTarget: target };
-    }
-    return { label: "Ran JS in agent browser" };
-  }
-  if (path.startsWith("/connections/browsers/")) return { label: "Agent browser action" };
-
-  if (path === "/connections") {
-    return { label: "Listed connections", connectionIconName: "connections" };
-  }
-  if (path.startsWith("/connections/")) {
-    const segments = path.split("/").slice(2); // [name, ...sub]
-    const name = segments[0];
-    const sub = segments.slice(1).join("/");
-    const icon = name;
-
-    // --- Gmail-specific labels (the connection has custom endpoints, not a proxy) ---
-    if (name === "gmail") {
-      if (sub === "send" && method === "POST") {
-        const body = curlBodyJson(cmd);
-        const to = typeof body?.to === "string" ? body.to : null;
-        return {
-          label: to ? `Sent email to ${trunc(to, 40)}` : "Sent email via Gmail",
-          connectionIconName: icon,
-        };
-      }
-      if (sub === "messages") {
-        const q = url.searchParams.get("q");
-        return {
-          label: q ? `Searched Gmail "${trunc(q, 30)}"` : "Listed Gmail messages",
-          connectionIconName: icon,
-        };
-      }
-      if (sub.startsWith("messages/")) {
-        return { label: "Read Gmail message", connectionIconName: icon };
-      }
-      if (sub === "instances") {
-        return { label: "Listed Gmail accounts", connectionIconName: icon };
-      }
-    }
-
-    // --- Google Calendar ---
-    if (name === "google-calendar") {
-      if (sub === "events") {
-        return { label: "Listed calendar events", connectionIconName: icon };
-      }
-      if (sub === "status") {
-        return { label: "Checked calendar connection", connectionIconName: icon };
-      }
-    }
-
-    // --- Proxy endpoints (Google Docs/Sheets, Notion, etc.) ---
-    if (sub.startsWith("proxy/")) {
-      const proxyPath = sub.slice("proxy/".length);
-      // Google Docs API
-      if (proxyPath.startsWith("docs/v1/documents")) {
-        if (method === "POST" && proxyPath.endsWith(":batchUpdate")) {
-          return { label: "Edited Google Doc", connectionIconName: icon };
-        }
-        if (method === "POST") {
-          return { label: "Created Google Doc", connectionIconName: icon };
-        }
-        return { label: "Read Google Doc", connectionIconName: icon };
-      }
-      // Drive API (used by google-docs for file listing + creation)
-      if (proxyPath.startsWith("drive/v3/files")) {
-        if (proxyPath.includes("/export")) {
-          return { label: "Exported Drive file", connectionIconName: icon };
-        }
-        if (method === "POST") {
-          return { label: "Created Drive file", connectionIconName: icon };
-        }
-        return { label: "Listed Drive files", connectionIconName: icon };
-      }
-      // Drive resumable/multipart upload
-      if (proxyPath.startsWith("upload/")) {
-        return { label: "Uploaded file to Drive", connectionIconName: icon };
-      }
-      // Google Sheets API
-      if (name === "google-sheets") {
-        if (proxyPath.endsWith(":append")) {
-          return { label: "Appended to sheet", connectionIconName: icon };
-        }
-        if (proxyPath.includes("/values/")) {
-          return {
-            label: method === "GET" ? "Read sheet values" : "Updated sheet values",
-            connectionIconName: icon,
-          };
-        }
-        return { label: "Sheets request", connectionIconName: icon };
-      }
-      // Generic proxy fallback — name the action by verb, not "Configured"
-      if (method === "POST") return { label: `Posted to ${name}`, connectionIconName: icon };
-      if (method === "PATCH" || method === "PUT") {
-        return { label: `Updated via ${name}`, connectionIconName: icon };
-      }
-      return { label: `Read from ${name}`, connectionIconName: icon };
-    }
-
-    // --- Catch-all for connection root + unrecognized subpaths ---
-    if (method === "DELETE") {
-      return { label: `Removed ${name} connection`, connectionIconName: icon };
-    }
-    // Root POST/PATCH/PUT on /connections/<id> is the actual "configure" action.
-    if (!sub && (method === "POST" || method === "PATCH" || method === "PUT")) {
-      return { label: `Configured ${name} connection`, connectionIconName: icon };
-    }
-    // Sub-path POST/PATCH/PUT is an action, not a configuration change.
-    if (method === "POST") return { label: `Posted to ${name}`, connectionIconName: icon };
-    if (method === "PATCH" || method === "PUT") {
-      return { label: `Updated via ${name}`, connectionIconName: icon };
-    }
-    return { label: `${name} connection`, connectionIconName: icon };
-  }
-
-  if (path === "/pipes") {
-    if (method === "POST") return { label: "Installed pipe" };
-    return { label: "Listed pipes" };
-  }
-  const pipeMatch = path.match(/^\/pipes\/([^/]+)(?:\/(.+))?$/);
-  if (pipeMatch) {
-    const name = pipeMatch[1];
-    const sub = pipeMatch[2];
-    if (sub === "executions") return { label: `${name}: recent runs` };
-    if (sub === "run" || method === "POST") return { label: `Ran pipe ${name}` };
-    if (method === "PATCH" || method === "PUT") return { label: `Configured pipe ${name}` };
-    if (method === "DELETE") return { label: `Removed pipe ${name}` };
-    return { label: `Pipe ${name}` };
-  }
-
-  if (path === "/export") return { label: "Exported video" };
-  if (path === "/health") return { label: "Health check" };
-  if (path === "/list-monitors") return { label: "Listed monitors" };
-  if (path === "/list-audio-devices") return { label: "Listed audio devices" };
-  if (path === "/tags") return { label: "Listed tags" };
-
-  return { label: `${method} ${path}` };
-}
-
 function extractAppFromToolCall(toolCall: ToolCall): string | undefined {
   if (toolCall.toolName === "bash") {
     return classifyCurl(String(toolCall.args?.command ?? ""))?.appName;
@@ -1289,104 +919,6 @@ function bashToolDetailsPresentation(toolCall: ToolCall): BashToolDetailsPresent
     rawCommand: command,
     rawResult: toolCall.result,
   };
-}
-
-function endpointFamily(path: string): string {
-  if (path === "/memories" || path.startsWith("/memories/")) return "Memory";
-  if (path === "/search") return "Screen search";
-  if (path === "/activity-summary") return "Activity";
-  if (path === "/raw_sql") return "Database";
-  if (path.startsWith("/connections/")) {
-    // Narrow the chip to the action surface, not just "Connection", so the AI's
-    // user-visible card matches the verb in the title (Sent email → EMAIL).
-    const segments = path.split("/").slice(2);
-    const name = segments[0];
-    const sub = segments.slice(1).join("/");
-    if (name === "gmail" && sub === "send") return "Email";
-    if (name === "gmail") return "Gmail";
-    if (name === "google-calendar") return "Calendar";
-    if (name === "google-docs") return "Doc";
-    if (name === "google-sheets") return "Sheet";
-    if (name === "slack") return "Slack";
-    if (name === "notion") return "Notion";
-    if (name === "telegram") return "Telegram";
-    if (name === "discord") return "Discord";
-    return "Connection";
-  }
-  if (path.startsWith("/meetings")) return "Meetings";
-  if (path.startsWith("/speakers")) return "Speakers";
-  if (path.startsWith("/pipes")) return "Pipes";
-  return "Screenpipe";
-}
-
-function parseToolResultJson(result: string | undefined): any | null {
-  if (!result?.trim()) return null;
-  try {
-    return JSON.parse(result);
-  } catch {
-    return null;
-  }
-}
-
-function summarizeToolResult(result: string | undefined, family: string): string | undefined {
-  const json = parseToolResultJson(result);
-  if (!json) return result?.trim() ? trunc(result.trim().replace(/\s+/g, " "), 120) : undefined;
-
-  // Connection-specific successes: read the actual response shape so the
-  // summary reflects what just happened ("Email sent", "Doc created") instead
-  // of the generic "JSON response returned" fallback.
-  if (family.startsWith("/connections/")) {
-    if (family === "/connections/gmail/send" && (json?.data?.id || json?.id || json?.threadId)) {
-      return "Email sent";
-    }
-    if (family.startsWith("/connections/google-docs/proxy/docs/v1/documents")) {
-      if (family.endsWith(":batchUpdate")) return "Document updated";
-      if (json?.documentId) return "Document created";
-    }
-    if (family.startsWith("/connections/google-docs/proxy/drive/v3/files") && json?.id) {
-      return json?.mimeType?.includes("spreadsheet") ? "Spreadsheet created" : "Drive file created";
-    }
-    if (family.startsWith("/connections/google-docs/proxy/upload/drive/v3/files") && json?.id) {
-      return "File uploaded";
-    }
-    if (family.endsWith(":append") && json?.updates?.updatedCells) {
-      return `Appended ${json.updates.updatedCells} cell${json.updates.updatedCells === 1 ? "" : "s"}`;
-    }
-    if (typeof json?.error === "string") return trunc(json.error, 120);
-  }
-
-  const noun = family === "/memories" ? "memories"
-    : family === "/search" ? "results"
-    : family.startsWith("/meetings") ? "meetings"
-    : family.startsWith("/connections") ? "items"
-    : "items";
-
-  if (Array.isArray(json)) return json.length === 0 ? `No ${noun} returned` : `${json.length} ${noun} returned`;
-  if (Array.isArray(json.data)) {
-    const total = typeof json.pagination?.total === "number" ? json.pagination.total : json.data.length;
-    return total === 0 ? `No ${noun} found` : `${total} ${noun} found`;
-  }
-  if (Array.isArray(json.search_results)) {
-    return json.search_results.length === 0 ? "No web sources returned" : `${json.search_results.length} web sources returned`;
-  }
-  if (Array.isArray(json.choices)) return `${json.choices.length} response${json.choices.length === 1 ? "" : "s"} returned`;
-  if (typeof json.success === "boolean") return json.success ? "Request succeeded" : "Request did not succeed";
-  if (typeof json.status === "string") return `Status: ${json.status}`;
-  return "JSON response returned";
-}
-
-function formatToolResult(result: string | undefined): string | undefined {
-  if (!result) return undefined;
-  const json = parseToolResultJson(result);
-  if (!json) return result;
-  return JSON.stringify(json, null, 2);
-}
-
-function sanitizeCommand(command: string): string {
-  return command
-    .replace(/\s-H\s+['"]Authorization:\s*Bearer\s+[^'"]+['"]/g, " -H \"Authorization: Bearer …\"")
-    .replace(/\s+/g, " ")
-    .trim();
 }
 
 function BashToolDetails({ toolCall }: { toolCall: ToolCall }) {
